LayoutTests/platform/gtk/media/media-source/media-source-resize-expected.txt

This tests the ability of the SourceBuffer to reset the parser after an abort(). A SourceBuffer in this state should be able to accept a new initialization segment or a new media segment.

RUN(video.src = URL.createObjectURL(source))
EVENT(sourceopen)
RUN(source.duration = loader.duration())
RUN(sourceBuffer = source.addSourceBuffer(loader.type()))
RUN(sourceBuffer.appendBuffer(loader.initSegment()))
EVENT(resize)
EXPECTED (video.videoWidth == '640') OK
EXPECTED (video.videoHeight == '480') OK
END OF TEST
